Include action buttons in actions component
