+++ src/components/atoms/button/index.jsx
@@ -1,6 +1,34 @@
-import Button from './regularButton';
-import CopyButton from './copyButton';
-import CodepenButton from './codepenButton';
-import ShareButton from './shareButton';
+import PropTypes from 'typedefs/proptypes';
+import { combineClassNames } from 'utils';
 
-export { Button, CopyButton, CodepenButton, ShareButton };
+const propTypes = {
+  onClick: PropTypes.func,
+  className: PropTypes.string,
+  children: PropTypes.oneOfType([
+    PropTypes.node,
+    PropTypes.arrayOf(PropTypes.node),
+  ]),
+  rest: PropTypes.any,
+};
+
+/**
+ * Generic button component.
+ */
+const Button = ({
+  onClick,
+  className = '',
+  children,
+  ...rest // Needs props to have accessible name if only icon etc.
+}) => (
+  <button
+    className={combineClassNames`btn fs-md ${className}`}
+    onClick={onClick}
+    {...rest}
+  >
+    {children}
+  </button>
+);
+
+Button.propTypes = propTypes;
+
+export default Button;
