--- src/components/atoms/button/regularButton/_index.scss
@@ -1,83 +1,0 @@
-.btn, a.btn {
-  display: inline-block;
-  padding: 0.625rem 0.875rem;
-  margin: 0 0.5rem;
-  font-weight: 500;
-  -webkit-font-smoothing: antialiased;
-  line-height: 1.5rem;
-  cursor: pointer;
-  border: none;
-  border-radius: var(--br-round);
-  transition: 0.3s ease all;
-
-  &:hover, &:focus {
-    animation: none;
-    outline: none;
-  }
-
-  // Utility for btn elements that look like actions
-  &.action-btn {
-    background: transparent;
-    color: var(--clr-txt-150);
-
-    &:hover, &:focus {
-      background: var(--clr-el-01dp);
-      color: var(--clr-primary-100);
-    }
-  }
-
-  &.outline-btn {
-    background: transparent;
-    border: 2px solid var(--clr-primary-100);
-    color: var(--clr-primary-050);
-
-    &:hover, &:focus {
-      background: var(--clr-primary-200);
-      border: 2px solid var(--clr-primary-200);
-      color: var(--clr-txt-200);
-    }
-  }
-
-  &.active {
-    animation-name: active;
-    animation-duration: 0.5s;
-    animation-timing-function: cubic-bezier(.4,0,.2,1);
-    transition: background 0.25s ease;
-
-    &:before {
-      opacity: 0;
-      animation-name: active-before;
-      animation-duration: 0.65s;
-      animation-timing-function: ease-in-out;
-    }
-  }
-}
-
-@keyframes active {
-  0%, 100% {
-    transform: scale(1);
-  }
-  20% {
-    transform: scale(1.15) rotate(3deg);
-  }
-  40% {
-    transform: scale(.94);
-  }
-  60% {
-    transform: scale(.98) rotate(-3deg);
-  }
-  80% {
-    transform: scale(1.08);
-  }
-  99% {
-  }
-}
-
-@keyframes active-before {
-  0%, 100% {
-    opacity: 0;
-  }
-  30%, 75% {
-    opacity: 1.0;
-  }
-}
