+++ src/components/molecules/actions/_index.scss
@@ -17,5 +17,48 @@
       vertical-align: bottom;
       width: 100%;
     }
+
+    &.active {
+      animation-name: active;
+      animation-duration: 0.5s;
+      animation-timing-function: cubic-bezier(.4,0,.2,1);
+      transition: background 0.25s ease;
+
+      &:before {
+        opacity: 0;
+        animation-name: active-before;
+        animation-duration: 0.65s;
+        animation-timing-function: ease-in-out;
+      }
+    }
+  }
+}
+
+@keyframes active {
+  0%, 100% {
+    transform: scale(1);
+  }
+  20% {
+    transform: scale(1.15) rotate(3deg);
+  }
+  40% {
+    transform: scale(.94);
+  }
+  60% {
+    transform: scale(.98) rotate(-3deg);
+  }
+  80% {
+    transform: scale(1.08);
+  }
+  99% {
+  }
+}
+
+@keyframes active-before {
+  0%, 100% {
+    opacity: 0;
+  }
+  30%, 75% {
+    opacity: 1.0;
   }
 }
