+++ src/components/molecules/cookieConsentPopup/index.jsx
@@ -1,7 +1,6 @@
 import { useState, useEffect } from 'react';
-import PropTypes from 'typedefs/proptypes';
 import Link from 'next/link';
-import { Button } from 'components/atoms/button';
+import Button from 'components/atoms/button';
 import { useShellDispatch } from 'state/shell';
 import literals from 'lang/en/client/cookieConsent';
